# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service for managing aircraft, flight, and voucher assignments",
        "title": "Voucher Seat Assignment API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
        "/aircraft": {
            "get": {
                "description": "Get all aircraft",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aircraft"
                ],
                "summary": "List all aircraft",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Aircraft"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new aircraft with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aircraft"
                ],
                "summary": "Create a new aircraft",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Aircraft"
                        }
                    }
                }
            }
        },
        "/aircraft/{id}": {
            "get": {
                "description": "Get aircraft details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aircraft"
                ],
                "summary": "Get an aircraft by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Aircraft ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aircraft"
                        }
                    }
                }
            },
            "put": {
                "description": "Update aircraft details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aircraft"
                ],
                "summary": "Update an aircraft",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Aircraft ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aircraft"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete aircraft by ID",
                "tags": [
                    "aircraft"
                ],
                "summary": "Delete an aircraft",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Aircraft ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/flights": {
            "get": {
                "description": "Get all flights",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flights"
                ],
                "summary": "List all flights",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Flight"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new flight with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flights"
                ],
                "summary": "Create a new flight",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Flight"
                        }
                    }
                }
            }
        },
        "/flights/{id}": {
            "get": {
                "description": "Get flight details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flights"
                ],
                "summary": "Get a flight by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Flight ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Flight"
                        }
                    }
                }
            },
            "put": {
                "description": "Update flight details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flights"
                ],
                "summary": "Update a flight",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Flight ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Flight"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete flight by ID",
                "tags": [
                    "flights"
                ],
                "summary": "Delete a flight",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Flight ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/vouchers": {
            "get": {
                "description": "Get a list of all vouchers with their associated flight information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vouchers"
                ],
                "summary": "List all vouchers",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved vouchers list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Voucher"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new voucher with crew information and seat assignments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vouchers"
                ],
                "summary": "Create a new voucher",
                "parameters": [
                    {
                        "description": "Voucher information",
                        "name": "voucher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Voucher"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created voucher",
                        "schema": {
                            "$ref": "#/definitions/models.Voucher"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/vouchers/{id}": {
            "get": {
                "description": "Get detailed voucher information including crew details and seat assignments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vouchers"
                ],
                "summary": "Get a voucher by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Voucher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved voucher",
                        "schema": {
                            "$ref": "#/definitions/models.Voucher"
                        }
                    },
                    "404": {
                        "description": "Voucher not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update voucher details including crew information and seat assignments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vouchers"
                ],
                "summary": "Update a voucher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Voucher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated voucher information",
                        "name": "voucher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Voucher"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated voucher",
                        "schema": {
                            "$ref": "#/definitions/models.Voucher"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Voucher not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a voucher and its associated data",
                "tags": [
                    "vouchers"
                ],
                "summary": "Delete a voucher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Voucher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted"
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Aircraft": {
            "type": "object",
            "properties": {
                "aircraft_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "num_rows": {
                    "type": "integer"
                },
                "registration_number": {
                    "type": "string"
                },
                "seats_per_row": {
                    "description": "Comma-separated seat letters, e.g. \"A,C,D,F\"",
                    "type": "string"
                }
            }
        },
        "models.Flight": {
            "type": "object",
            "properties": {
                "aircraft": {
                    "$ref": "#/definitions/models.Aircraft"
                },
                "aircraft_id": {
                    "type": "integer"
                },
                "aircraft_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "flight_date": {
                    "type": "string"
                },
                "flight_number": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "models.Voucher": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "crew_id": {
                    "type": "string"
                },
                "crew_name": {
                    "type": "string"
                },
                "flight": {
                    "$ref": "#/definitions/models.Flight"
                },
                "flight_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "seat1": {
                    "type": "string"
                },
                "seat2": {
                    "type": "string"
                },
                "seat3": {
                    "type": "string"
                }
            }
        }
    }
}